+++ yaralyzer/__init__.py
@@ -4,7 +4,7 @@ import yara as python_yara
 from yaralyzer.config import YaralyzerConfig
 from yaralyzer.output.console import console
 from yaralyzer.output.file_export import export_json, invoke_rich_export
-from yaralyzer.util.argument_parser import parser, parse_arguments
+from yaralyzer.util.argument_parser import parser
 from yaralyzer.util.constants import PDFALYZER_MSG_TXT
 from yaralyzer.util.exceptions import print_fatal_error_and_exit
 from yaralyzer.util.logging import invocation_txt, log_console
@@ -12,7 +12,7 @@ from yaralyzer.yara.error import yara_error_msg
 from yaralyzer.yara.yara_rule_builder import HEX, REGEX
 from yaralyzer.yaralyzer import Yaralyzer
 
-YaralyzerConfig.init(parser, parse_arguments)
+YaralyzerConfig.init(parser)
 
 
 def yaralyze():